Renaming utils folder and references to avoid stomping on wp.utils
